[Refactor] 패키지 구조 변경 - GateWayModule 추가 및 큐 돌리던중 discconect 감지 #37
