+++ src/domain/game/game.module.ts
@@ -2,13 +2,12 @@ import { Module } from '@nestjs/common';
 import { GameService } from './game.service';
 import { FactoryModule } from '../factory/factory.module';
 import { GameController } from './game.controller';
-import { QueueService } from '../queue/queue.service';
-import { GameGateWay } from './game.gateway';
+import { GatewayModule } from '../gateway/gateway.module';
 import { QueueModule } from '../queue/queue.module';
 
 @Module({
-  imports: [FactoryModule, QueueModule],
-  providers: [GameService, GameGateWay],
+  imports: [FactoryModule, GatewayModule, QueueModule],
+  providers: [GameService],
   exports: [GameService],
   controllers: [GameController],
 })
