+++ src/domain/user/user.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
-import { PostGatewayUserDto } from './post.gateway.users.dto';
+import { PostGatewayUserDto } from './dto/post.gateway.users.dto';
 
 @Injectable()
 export class UserRepository {
